requirements.txt
flask~=2.3.2
flask-smorest
python-dotenv
marshmallow~=3.19.0
sqlalchemy~=2.0.15
flask-sqlalchemy
gunicorn
flask-jwt-extended
flask-migrate
passlib~=1.7.4
python-dotenv
psycopg2-binary
requests
Faker~=18.10.1
socketio~=0.2.1
PyYAML~=6.0
pytest~=7.3.1

aiohttp~=3.8.4
alembic~=1.11.1
PyJWT~=2.7.0
python-dateutil~=2.8.2
Cerberus~=1.3.4